requirements.txt
streamlit
nltk
spacy
scikit-learn
PyPDF2
python-docx
